muppetsitem test added

+++ src/main/java/film_memorabilia/MuppetsItem.java
@@ -5,7 +5,7 @@ import enums.ItemType;
 public class MuppetsItem extends Item {
     private ItemType type;
 
-    public MuppetsItem(ItemType type, String itemDescription, int buyPrice, int shipping, int marketValue) {
+    public MuppetsItem(ItemType type, String itemDescription, double buyPrice, double shipping, double marketValue) {
         super(itemDescription, buyPrice, shipping, marketValue);
         this.type = type;
     }
